Allow boolean proxy=False to override PROXY
I needed a simple way to disable the default proxy list for certain API calls. Setting proxy=None would still invoke the PROXY list because each endpoint is passing proxy=None to NBAHTTP().send_api_request() even when proxy is not explicitly passed.

If you catch this information in the endpoints, it would still be possible to use a mutable object to detect when a user wants to forcefully override the PROXY default (useful for debugging). See https://github.com/swar/nba_api/pull/162 as an example (you would need to compare self.send_api_request.__defaults__[n]: within each endpoint, where n is the position of 'proxy'). As it stands now, empty user input (i.e. default) is masked so we cannot determine default from override behaviors. As that solution involves editing every endpoint, I'll leave that up to you to decide if it is worth it.